scsi: pm8001: Use sas_execute_ssp_tmf()

+++ drivers/scsi/pm8001/pm8001_hwi.c
@@ -4648,7 +4648,7 @@ int pm8001_chip_ssp_tm_req(struct pm8001_hba_info *pm8001_ha,
 
 	memset(&sspTMCmd, 0, sizeof(sspTMCmd));
 	sspTMCmd.device_id = cpu_to_le32(pm8001_dev->device_id);
-	sspTMCmd.relate_tag = cpu_to_le32((u32)tmf->tag);
+	sspTMCmd.relate_tag = cpu_to_le32(tmf->tag);
 	sspTMCmd.tmf = cpu_to_le32(tmf->tmf);
 	memcpy(sspTMCmd.lun, task->ssp_task.LUN, 8);
 	sspTMCmd.tag = cpu_to_le32(ccb->ccb_tag);
